refactor(architecture): move gateways from domain/model to domain/usecase context packages
- Move OrderGateway to order/gateway
- Move ProductGateway and ProductCategoryGateway to product/gateway
- Move SupplierGateway, CompanyGateway, CustomerGateway, AddressGateway, ContactGateway, EmployeeGateway to party/gateway
- Move WarehouseGateway, WarehouseInventoryGateway, InventoryMovementGateway to inventory/gateway
- Move ShoppingCartGateway to cart/gateway
- Move ShippingDetailGateway to shipping/gateway
- Move CustomerGateway to report/gateway
- Update all imports across use cases, services and tests
