+++ domain/usecase/src/main/java/com/arka/cart/AddItemToShoppingCartUseCase.java
@@ -2,10 +2,10 @@ package com.arka.cart;
 
 import com.arka.cart.dto.AddItemShoppingCartIn;
 import com.arka.cart.dto.ShoppingCartOut;
+import com.arka.cart.mapper.ShoppingCartMapperImpl;
 import com.arka.enums.ShoppingCartStatus;
-import com.arka.gateway.ShoppingCartGateway;
+import com.arka.cart.gateway.ShoppingCartGateway;
 import com.arka.cart.mapper.ShoppingCartMapper;
-import com.arka.mapper.ShoppingCartMapperImpl;
 import com.arka.model.cart.ShoppingCart;
 import com.arka.model.product.Product;
 import com.arka.product.service.ProductService;
