+++ domain/usecase/src/main/java/com/arka/inventory/gateway/WarehouseGateway.java
@@ -1,4 +1,4 @@
-package com.arka.gateway;
+package com.arka.inventory.gateway;
 
 import com.arka.model.Warehouse;
