+++ domain/usecase/src/main/java/com/arka/inventory/service/WarehouseService.java
@@ -1,7 +1,7 @@
 package com.arka.inventory.service;
 
 import com.arka.exceptions.NotFoundException;
-import com.arka.gateway.WarehouseGateway;
+import com.arka.inventory.gateway.WarehouseGateway;
 import com.arka.model.Warehouse;
 import lombok.RequiredArgsConstructor;
