+++ domain/usecase/src/main/java/com/arka/order/CreateOrderUseCase.java
@@ -2,7 +2,7 @@ package com.arka.order;
 
 import com.arka.order.dto.CreateOrderIn;
 import com.arka.order.dto.CreateOrderOut;
-import com.arka.gateway.order.OrderGateway;
+import com.arka.order.gateway.OrderGateway;
 import com.arka.order.mapper.OrderMapper;
 import com.arka.model.Company;
 import com.arka.model.order.Order;
