+++ domain/usecase/src/main/java/com/arka/order/gateway/OrderGateway.java
@@ -1,4 +1,4 @@
-package com.arka.gateway.order;
+package com.arka.order.gateway;
 
 import com.arka.model.order.Order;
