+++ domain/usecase/src/main/java/com/arka/party/ListSuppliersByCategoryUseCase.java
@@ -2,9 +2,9 @@ package com.arka.party;
 
 import com.arka.party.dto.CompanyOut;
 import com.arka.party.mapper.CompanyMapper;
-import com.arka.mapper.CompanyMapperImpl;
 import com.arka.model.Company;
-import com.arka.gateway.party.SupplierGateway;
+import com.arka.party.gateway.SupplierGateway;
+import com.arka.party.mapper.CompanyMapperImpl;
 import com.arka.product.service.ProductCategoryService;
 import lombok.RequiredArgsConstructor;
