+++ domain/usecase/src/main/java/com/arka/party/gateway/EmployeeGateway.java
@@ -1,4 +1,4 @@
-package com.arka.gateway;
+package com.arka.party.gateway;
 
 import com.arka.model.Employee;
