+++ domain/usecase/src/main/java/com/arka/party/service/AddressService.java
@@ -1,7 +1,7 @@
 package com.arka.party.service;
 
 import com.arka.exceptions.NotFoundException;
-import com.arka.gateway.AddressGateway;
+import com.arka.party.gateway.AddressGateway;
 import com.arka.model.information.Address;
 import lombok.RequiredArgsConstructor;
