+++ domain/usecase/src/main/java/com/arka/party/service/EmployeeService.java
@@ -1,7 +1,7 @@
 package com.arka.party.service;
 
 import com.arka.exceptions.NotFoundException;
-import com.arka.gateway.EmployeeGateway;
+import com.arka.party.gateway.EmployeeGateway;
 import com.arka.model.Employee;
 import lombok.RequiredArgsConstructor;
