+++ domain/usecase/src/main/java/com/arka/product/CreateProductUseCase.java
@@ -2,11 +2,11 @@ package com.arka.product;
 
 import com.arka.product.dto.CreateProductIn;
 import com.arka.product.dto.CreateProductOut;
-import com.arka.gateway.product.ProductGateway;
+import com.arka.product.gateway.ProductGateway;
 import com.arka.product.mapper.CreateProductOutMapper;
-import com.arka.mapper.CreateProductOutMapperImpl;
 import com.arka.model.product.Product;
 import com.arka.model.product.ProductCategory;
+import com.arka.product.mapper.CreateProductOutMapperImpl;
 import com.arka.product.service.ProductCategoryService;
 import com.arka.util.NullValidator;
 import lombok.RequiredArgsConstructor;
