+++ domain/usecase/src/main/java/com/arka/product/service/ProductService.java
@@ -3,7 +3,7 @@ package com.arka.product.service;
 import com.arka.report.dto.ProductSalesReportOut;
 import com.arka.report.dto.InstantDateRange;
 import com.arka.exceptions.NotFoundException;
-import com.arka.gateway.product.ProductGateway;
+import com.arka.product.gateway.ProductGateway;
 import com.arka.model.product.Product;
 import com.arka.util.DateRangeTypeParser;
 import com.arka.util.DateRangeValidator;
