+++ domain/usecase/src/main/java/com/arka/shipping/RegisterShippingDetailsUseCase.java
@@ -2,14 +2,14 @@ package com.arka.shipping;
 
 import com.arka.shipping.dto.CreateShippingDetailIn;
 import com.arka.shipping.dto.ShippingDetailOut;
-import com.arka.gateway.ShippingDetailGateway;
+import com.arka.shipping.gateway.ShippingDetailGateway;
 import com.arka.shipping.mapper.ShippingDetailMapper;
-import com.arka.mapper.ShippingDetailMapperImpl;
 import com.arka.model.ShippingDetail;
 import com.arka.model.information.Address;
 import com.arka.model.order.Order;
 import com.arka.party.service.AddressService;
 import com.arka.order.service.OrderService;
+import com.arka.shipping.mapper.ShippingDetailMapperImpl;
 import com.arka.util.NullValidator;
 import lombok.RequiredArgsConstructor;
