+++ domain/usecase/src/test/java/com/arka/usecase/order/ModifyOrderUseCaseTest.java
@@ -4,7 +4,7 @@ import com.arka.order.ModifyOrderUseCase;
 import com.arka.order.dto.UpdateOrderIn;
 import com.arka.enums.CompanyRelationType;
 import com.arka.enums.OrderType;
-import com.arka.gateway.order.OrderGateway;
+import com.arka.order.gateway.OrderGateway;
 import com.arka.order.mapper.OrderMapper;
 import com.arka.model.Company;
 import com.arka.model.order.Order;
