+++ domain/usecase/src/test/java/com/arka/usecase/party/CreateSupplierUseCaseTest.java
@@ -3,7 +3,7 @@ package com.arka.usecase.party;
 import com.arka.party.CreateSupplierUseCase;
 import com.arka.party.dto.CreateSupplierIn;
 import com.arka.enums.CompanyRelationType;
-import com.arka.gateway.party.SupplierGateway;
+import com.arka.party.gateway.SupplierGateway;
 import com.arka.model.information.Contact;
 import com.arka.model.product.ProductCategory;
 import com.arka.party.service.ContactService;
